+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tcswap/browser
 
+## 4.1.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tcswap/helpers@4.5.18
+  - @tcswap/core@4.1.20
+  - @tcswap/plugins@4.2.18
+  - @tcswap/wallets@4.3.20
+
 ## 4.1.19
 
 ### Patch Changes