+++ packages/server/CHANGELOG.md
@@ -1,5 +1,15 @@
 <!-- Modifications © 2025 Horizontal Systems. -->
 
+## 4.0.62
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tcswap/helpers@4.5.18
+  - @tcswap/toolboxes@4.3.18
+  - @tcswap/tokens@4.1.5
+  - @tcswap/wallet-keystore@4.1.18
+
 ## 4.0.61
 
 ### Patch Changes