+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 <!-- Modifications © 2025 Horizontal Systems. -->
 
+## 0.3.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tcswap/helpers@4.5.18
+  - @tcswap/sdk@4.2.20
+  - @tcswap/wallet-keystore@4.1.18
+  - @tcswap/wallets@4.3.20
+
 ## 0.3.19
 
 ### Patch Changes